cleanup file handling functions and remove FileUploadTestScreen
