+++ apps/app/types/navigation.ts
@@ -81,7 +81,6 @@ export type RootStackParamList = {
   Login: LoginParams | undefined;
   LogoutInProgress: undefined;
   EncryptDecryptImageTest: undefined;
-  FileUploadTest: undefined;
   TestLibsodium: undefined;
   AccountSettings: AccountSettingsParams | undefined;
   AccountSettingsProfile: undefined; // on phones
